pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["gitme", "gitme.*"]
namespaces = true

[project]
name = "GitMe"
version = "0.1.0"
description = "A library for summarizing data about Your GitHub repositories using various generative AI service provision options"
license = {file = "LICENSE"}
dependencies = [
    "google-generativeai==0.6.0",
    "pygithub==2.3.0",
    "pydantic==2.7.0",
    "tenacity==8.3.0"
]
